Update to Version 1.2
fixed noexcept so that it now gets used in Visual Studio 2015 or later
but still doesn't in earlier versions as it's not supported. it also now
uses its own unique macro identifier.

added ability to load a buffer from char* instead or requiring a vector
of char.

added ability to manually set and get the maximum number of voices
allowed. setting this immediately resizes the vector that holds all of
the sounds.

added ability to play a sound at a specific location using only a 2D
co-ordinate. that is, a depth of effectively zero. however, the depth is
actually a very low positive value so that all positions are definitely
in front of the camera (assuming that +z is away from the camera).