Add Extra Music Control
added extra controls for music to allow pausing and resuming the
(current) music instead of only playing and stopping.
also added ability to get the position (in sf::Time) of the current
music. if the current music is stopped, the time returned is zero. note
that no checks are made to make sure that the music actually exists, as
with getCurrentMusicStatus.