Update to v1.4
added ability to copy a sound fx buffer from an external sound buffer directly; presumes full volume but can be modified afterwards.

added ability to get current music pause state; that is, whether the current music's state is current paused or not - returns a boolean.

added ability to 'restart' the current music from the beginning. this applies when the music is playing, stopped or paused.

updated year and version number.