update to v1.6
updates to version 1.6, bringing all the code into a single header file.

this removes the 3 separate files (Common.hpp, Control.hpp and Control.cpp) and places all relevant code from them into what was just the link header, SfmlSoundSystem.hpp, that is in the root folder. This file is now the entire code.